+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Implement tool tips and 'required' asterisk for some forms.
 * Replace `type` field with `Record payments separately` checkbox.
 * Rename 'Cash sale' to just 'Sale', 'Cash purchase' to just 'Purchase'
+* Implement onboarding.
 
 
 # 1.0.1 - 20201120